+++ lib/constants.rb
@@ -17,9 +17,14 @@ USER_DIR = "~/.kindlemail"
 STORAGE_DIR = USER_DIR + "/.storage"
 EMAIL_CONF_FILE = File.expand_path(USER_DIR + "/.email_conf")
 USER_CONF_FILE = File.expand_path(USER_DIR + "/.kindlemail")
-VERSION = "0.2.5"
+
+version_file = File.expand_path(File.join(File.dirname(__FILE__), "../", "VERSION"))
+File.open(version_file, 'r') do |file| 
+  APP_VERSION = file.gets
+end
+
 HOMEPAGE = "https://github.com/djhworld/kindlemail"
 AUTHOR = "djhworld"
-VERSION_STRING = "kindlemail #{VERSION}. Written by #{AUTHOR}. #{HOMEPAGE}"
+VERSION_STRING = "kindlemail #{APP_VERSION}. Written by #{AUTHOR}. #{HOMEPAGE}"
 FILE_STORE = File.expand_path(STORAGE_DIR + "/sent_files.history")